AMI provisioner doesn't support SSH password
WTF.  So the only way to auth to an instance is with an ssh key.

Also adding the ec2-user to the wheel group, as this is the group that
generally gets the sudo privileges.  The group has to be set before
the user logs in, or it won't be in effect for that login session.

Conflicts:
	packer.json